Change amount to BigInteger type

+++ src/main/java/me/vante/bitkassa/model/Invoice.java
@@ -2,6 +2,8 @@ package me.vante.bitkassa.model;
 
 import lombok.Data;
 
+import java.math.BigInteger;
+
 
 /**
  * Created by robbertcoeckelbergh on 3/11/16.
@@ -12,7 +14,7 @@ public class Invoice {
     private String payment_id;
     private String payment_url;
     private String address;
-    private int amount;
+    private BigInteger amount;
     private String bitcoin_url;
     private int expire;
     private Boolean success;
